ADDED API KEY MANAGEMENT

+++ backend/src/module/logo/logo.controller.ts
@@ -15,7 +15,7 @@ export class LogoController {
     return this.logoService.addLogo(user_id, logo_path);
   }
   @Delete()
-  deleteLogo(@Body("logo_path") logo_path: string) {
-    return this.logoService.deleteLogo(logo_path)
+  deleteLogo(@Body("logo_path") logo_path: string, @Body("user_id") user_id: string, @Body("logo_id") logo_id: number) {
+    return this.logoService.deleteLogo({ logo_path, user_id, logo_id })
   }
 }
